requirements.txt
Pillow==2.4.0
ipdb==0.8
ipython==2.0.0
lastfmapi==0.1
python-memcached==1.53
wsgiref==0.1.2
django>=1.6
octopus-http==0.6.4
